--- a/cmd/frontend/graphqlbackend/repository_mirror.go
+++ b/cmd/frontend/graphqlbackend/repository_mirror.go
@@ -32,11 +32,11 @@ type repositoryMirrorInfoResolver struct {
 
 	// memoize the gitserver RepoInfo call
 	repoInfoOnce     sync.Once
-	repoInfoResponse *protocol.RepoInfoResponse
+	repoInfoResponse *protocol.RepoInfo
 	repoInfoErr      error
 }
 
-func (r *repositoryMirrorInfoResolver) gitserverRepoInfo(ctx context.Context) (*protocol.RepoInfoResponse, error) {
+func (r *repositoryMirrorInfoResolver) gitserverRepoInfo(ctx context.Context) (*protocol.RepoInfo, error) {
 	r.repoInfoOnce.Do(func() {
 		resp, err := gitserver.DefaultClient.RepoInfo(ctx, r.repository.repo.Name)
 		r.repoInfoResponse, r.repoInfoErr = resp.Results[r.repository.repo.Name], err