--- a/pkg/gitserver/client.go
+++ b/pkg/gitserver/client.go
@@ -19,6 +19,7 @@ import (
 	"sync/atomic"
 	"time"
 
+	"github.com/hashicorp/go-multierror"
 	"github.com/neelance/parallel"
 	"github.com/opentracing-contrib/go-stdlib/nethttp"
 	opentracing "github.com/opentracing/opentracing-go"
@@ -558,21 +559,70 @@ func (c *Client) IsRepoCloned(ctx context.Context, repo api.RepoName) (bool, err
 // If multiple errors occurred, an incomplete result is returned along with a
 // *multierror.Error.
 func (c *Client) RepoInfo(ctx context.Context, repos ...api.RepoName) (*protocol.RepoInfoResponse, error) {
-	req := &protocol.RepoInfoRequest{
-		Repos: repos,
-	}
-	resp, err := c.httpPost(ctx, repos[0], "repos", req)
-	if err != nil {
-		return nil, err
-	}
-	defer resp.Body.Close()
-	if resp.StatusCode != http.StatusOK {
-		return nil, &url.Error{URL: resp.Request.URL.String(), Op: "RepoInfo", Err: fmt.Errorf("RepoInfo: http status %d", resp.StatusCode)}
+	shards := make(map[string]*protocol.RepoInfoRequest)
+
+	for _, r := range repos {
+		addr := c.addrForRepo(ctx, r)
+		shard := shards[addr]
+
+		if shard == nil {
+			shard = new(protocol.RepoInfoRequest)
+			shards[addr] = shard
+		}
+
+		shard.Repos = append(shard.Repos, r)
 	}
 
-	var info *protocol.RepoInfoResponse
-	err = json.NewDecoder(resp.Body).Decode(&info)
-	return info, err
+	type op struct {
+		req *protocol.RepoInfoRequest
+		res *protocol.RepoInfoResponse
+		err error
+	}
+
+	ch := make(chan op, len(shards))
+	for _, req := range shards {
+		go func(o op) {
+			var resp *http.Response
+			resp, o.err = c.httpPost(ctx, o.req.Repos[0], "repos", o.req)
+			if o.err != nil {
+				ch <- o
+				return
+			}
+
+			defer resp.Body.Close()
+			if resp.StatusCode != http.StatusOK {
+				o.err = &url.Error{
+					URL: resp.Request.URL.String(),
+					Op:  "RepoInfo",
+					Err: errors.Errorf("RepoInfo: http status %d", resp.StatusCode),
+				}
+			}
+
+			o.res = new(protocol.RepoInfoResponse)
+			o.err = json.NewDecoder(resp.Body).Decode(o.res)
+			ch <- o
+		}(op{req: req})
+	}
+
+	err := new(multierror.Error)
+	res := protocol.RepoInfoResponse{
+		Results: make(map[api.RepoName]*protocol.RepoInfo),
+	}
+
+	for i := 0; i < cap(ch); i++ {
+		o := <-ch
+
+		if o.err != nil {
+			err = multierror.Append(err, o.err)
+			continue
+		}
+
+		for repo, info := range o.res.Results {
+			res.Results[repo] = info
+		}
+	}
+
+	return &res, err.ErrorOrNil()
 }
 
 // Remove removes the repository clone from gitserver.