@slimsag 2019-04-03
I plan to update this + merge soon, but some other customer work became more important. @keegancsmith after I address @tsenart's comment (i.e. change to just a singular gitserver HTTP endpoint); any further comments?