@keegancsmith 2019-04-16
I'm keen to review the updated version. I agree with Tomas's feedback about simplifying the client API