@slimsag 2019-04-16
@keegancsmith @tsenart PTAL

I am positive you'll have more suggestions on how to make the code nicer (I can think of a few ways I may try tomorrow), but please review now.

If you want me to delay merging this until 3.4 please let me know.