@tsenart 2019-04-16
It seems that the last missing thing to agree on is whether the roll-out needs to be backwards compatible or not.