@slimsag 2019-04-16
@tsenart thanks for the feedback here & raising the important point of rollout. I've thought on this a bit today and:

1. I added back the deprecated gitserver `/repo` route, so that gitserver is backwards compat (can run old frontend with new gitserver).
2. I have not made new frontend compatible with old gitserver, I think it would be very tricky to do this correctly and if we wanted this a better approach would be rolling out the gitserver change in one version then the frontend change in the next (but this has the obvious drawback of 2 versions needed for making this change).
3. For Sourcegraph.com, rollout will be as simple as merging the gitserver renovate PR first since we have back compat (see #1 above), then the frontend renovate PR.
4. For customer Kubernetes deployments, there will be a small period where either gitserver or frontend could be up first. This period wouldn't last long, though, so I think it's OK. It would be exactly the amount of time it takes for the new gitserver to fully spin up -- which seems reasonable. In this timeframe, requests for code intel would fail.

Hopefully all of that sounds reasonable (if not I can revert before we release).